float instead of int

+++ webgnome_api/views/goods.py
@@ -421,7 +421,7 @@ class GOODSRequest(object):
         self.subset_size = None
         self.filename = filename
         self.outpath = outpath
-        self.tshift = int(tshift) #timezone shift retained for future use by webgnomeapi
+        self.tshift = float(tshift) #timezone shift retained for future use by webgnomeapi
         self._debug = _debug
         self._max_size = _max_size
         self._reconfirm_timeout = _reconfirm_timeout
